refactor(test): align pledge test suite with feature architecture
Migrate unit and integration tests for the `Pledge` domain to reflect
the new backend package-by-feature structure.

- Relocate tests to `pledge` package structure.
-  Update package declarations for Controller, Service, Mapper and
Entity tests.
- Update imports to reference the relocated `maintenance` and `pledge`
components (Entities, DTOs, Services).
- Update exception handling imports to `common` feature package.

Closes #35
