+++ src/test/java/com/unnamed/conectareparo/pledge/controller/PledgeControllerTest.java
@@ -1,13 +1,13 @@
-package com.unnamed.conectareparo.controller;
+package com.unnamed.conectareparo.pledge.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.unnamed.conectareparo.dto.PledgeDto;
-import com.unnamed.conectareparo.dto.PledgeResponseDto;
-import com.unnamed.conectareparo.dto.PledgeUpdateDto;
-import com.unnamed.conectareparo.entity.PledgeCategory;
-import com.unnamed.conectareparo.entity.PledgeStatus;
-import com.unnamed.conectareparo.exception.ResourceNotFoundException;
-import com.unnamed.conectareparo.service.PledgeService;
+import com.unnamed.conectareparo.pledge.dto.PledgeDto;
+import com.unnamed.conectareparo.pledge.dto.PledgeResponseDto;
+import com.unnamed.conectareparo.pledge.dto.PledgeUpdateDto;
+import com.unnamed.conectareparo.pledge.entity.PledgeCategory;
+import com.unnamed.conectareparo.pledge.entity.PledgeStatus;
+import com.unnamed.conectareparo.pledge.service.PledgeService;
+import com.unnamed.conectareparo.common.exception.ResourceNotFoundException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
