+++ src/test/java/com/unnamed/conectareparo/pledge/mapper/PledgeMapperTest.java
@@ -1,11 +1,11 @@
-package com.unnamed.conectareparo.mapper;
+package com.unnamed.conectareparo.pledge.mapper;
 
-import com.unnamed.conectareparo.dto.PledgeDto;
-import com.unnamed.conectareparo.dto.PledgeResponseDto;
-import com.unnamed.conectareparo.entity.Maintenance;
-import com.unnamed.conectareparo.entity.Pledge;
-import com.unnamed.conectareparo.entity.PledgeCategory;
-import com.unnamed.conectareparo.entity.PledgeStatus;
+import com.unnamed.conectareparo.maintenance.entity.Maintenance;
+import com.unnamed.conectareparo.pledge.dto.PledgeDto;
+import com.unnamed.conectareparo.pledge.dto.PledgeResponseDto;
+import com.unnamed.conectareparo.pledge.entity.Pledge;
+import com.unnamed.conectareparo.pledge.entity.PledgeCategory;
+import com.unnamed.conectareparo.pledge.entity.PledgeStatus;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
